 \name{DynamicDistribution-internal} 
        \alias{Con_Uniform} 
        \alias{Normal} 
        \alias{Chi_Square} 
        \alias{Student_t} 
        \alias{Exponential}
        \alias{Gamma_dis} 
        \alias{Beta_dis} 
        \alias{Laplace} 
        \alias{Logistic} 
        \alias{Lognormal}
        \alias{F_dis} 
        \alias{Pareto} 
        \alias{Cauchy} 
        \alias{Inverse_Gaussian} 
        \alias{Rayleigh}
        \alias{Dis_Uniform} 
        \alias{Bernoulli} 
        \alias{Binomial} 
        \alias{Hypergeometric} 
        \alias{Poisson}
        \alias{Geometric} 
        \alias{Negative_Binomial} 
        \alias{Logarithmic_Series}
        \title{Internal DynamicDistribution Functions} 
        \description{ 
          Internal mypackage functions 
        } 
        \details{ 
          These are not to be called by the user. 
        } 
        \keyword{ internal } 